pyproject.toml
[tool.poetry]
name = "bibliophilia"
version = "0.1.0"
description = ""
authors = ["BaDum-Tss","rafi2002", "Pechenuxa1"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
elasticsearch = "^8.10.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
